fix: rename module and package paths
