+++ packages/go/go.mod
@@ -1,3 +1,3 @@
-module github.com/ossf/osv-schema
+module github.com/ossf/osv-schema/packages/go
 
 go 1.22.7
